third_party/WebKit/LayoutTests/editing/deleting/delete-block-contents.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => {
    assert_selection(
        '<div contenteditable><span>|foo   bar             baz</span></div>',
        selection => {
            selection.document.execCommand('selectAll');
            selection.document.execCommand('delete');
        },
        '<div contenteditable>|<br></div>',
        'Delete after select all in SPAN');
    assert_selection(
        '<div contenteditable><span>^foo   bar             baz|</span></div>',
        'delete',
        '<div contenteditable>|<br></div>',
        'Delete all contens in SPAN');
}, 'Delete block');
</script>
